// packages/tracing/tests/generated/trace.ts
export interface TracesData {
  /**
   * An array of ResourceSpans.
   * For data coming from a single resource this array will typically contain
   * one element. Intermediary nodes that receive data from multiple origins
   * typically batch the data before forwarding further and in that case this
   * array will contain multiple elements.
   *
   * @generated from protobuf field: repeated opentelemetry.proto.trace.v1.ResourceSpans resource_spans = 1
   */
  resourceSpans: ResourceSpans[];
}
/**
 * A collection of ScopeSpans from a Resource.
 *
 * @generated from protobuf message opentelemetry.proto.trace.v1.ResourceSpans
 */
export interface ResourceSpans {
  /**
   * The resource for the spans in this message.
   * If this field is not set then no resource info is known.
   *
   * @generated from protobuf field: opentelemetry.proto.resource.v1.Resource resource = 1
   */
  resource?: Resource;
  /**
   * A list of ScopeSpans that originate from a resource.
   *
   * @generated from protobuf field: repeated opentelemetry.proto.trace.v1.ScopeSpans scope_spans = 2
   */
  scopeSpans: ScopeSpans[];
  /**
   * The Schema URL, if known. This is the identifier of the Schema that the resource data
   * is recorded in. Notably, the last part of the URL path is the version number of the
   * schema: http[s]://server[:port]/path/<version>. To learn more about Schema URL see
   * https://opentelemetry.io/docs/specs/otel/schemas/#schema-url
   * This schema_url applies to the data in the "resource" field. It does not apply
   * to the data in the "scope_spans" field which have their own schema_url field.
   *
   * @generated from protobuf field: string schema_url = 3
   */
  schemaUrl: string;
}
/**
 * A collection of Spans produced by an InstrumentationScope.
 *
 * @generated from protobuf message opentelemetry.proto.trace.v1.ScopeSpans
 */
export interface ScopeSpans {
  /**
   * The instrumentation scope information for the spans in this message.
   * Semantically when InstrumentationScope isn't set, it is equivalent with
   * an empty instrumentation scope name (unknown).
   *
   * @generated from protobuf field: opentelemetry.proto.common.v1.InstrumentationScope scope = 1
   */
  scope?: InstrumentationScope;
  /**
   * A list of Spans that originate from an instrumentation scope.
   *
   * @generated from protobuf field: repeated opentelemetry.proto.trace.v1.Span spans = 2
   */
  spans: Span[];
  /**
   * The Schema URL, if known. This is the identifier of the Schema that the span data
   * is recorded in. Notably, the last part of the URL path is the version number of the
   * schema: http[s]://server[:port]/path/<version>. To learn more about Schema URL see
   * https://opentelemetry.io/docs/specs/otel/schemas/#schema-url
   * This schema_url applies to the data in the "scope" field and all spans and span
   * events in the "spans" field.
   *
   * @generated from protobuf field: string schema_url = 3
   */
  schemaUrl: string;
}
// ...
class TracesData$Type extends MessageType<TracesData> {
  constructor() {
    super('opentelemetry.proto.trace.v1.TracesData', [
      { no: 1, name: 'resource_spans', kind: 'message', repeat: 2 /*RepeatType.UNPACKED*/, T: () => ResourceSpans },
    ]);
  }
  create(value?: PartialMessage<TracesData>): TracesData {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.resourceSpans = [];
    if (value !== undefined) reflectionMergePartial<TracesData>(this, message, value);
    return message;
  }
  internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: TracesData): TracesData {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated opentelemetry.proto.trace.v1.ResourceSpans resource_spans */ 1:
          message.resourceSpans.push(ResourceSpans.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw') throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          let d = reader.skip(wireType);
          if (u !== false) (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
      }
    }
    return message;
  }
  internalBinaryWrite(message: TracesData, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* repeated opentelemetry.proto.trace.v1.ResourceSpans resource_spans = 1; */
    for (let i = 0; i < message.resourceSpans.length; i++)
      ResourceSpans.internalBinaryWrite(message.resourceSpans[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.trace.v1.TracesData
 */
export const TracesData = new TracesData$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ResourceSpans$Type extends MessageType<ResourceSpans> {
  constructor() {
    super('opentelemetry.proto.trace.v1.ResourceSpans', [
      { no: 1, name: 'resource', kind: 'message', T: () => Resource },
      { no: 2, name: 'scope_spans', kind: 'message', repeat: 2 /*RepeatType.UNPACKED*/, T: () => ScopeSpans },
      { no: 3, name: 'schema_url', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ]);
  }
  create(value?: PartialMessage<ResourceSpans>): ResourceSpans {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.scopeSpans = [];
    message.schemaUrl = '';
    if (value !== undefined) reflectionMergePartial<ResourceSpans>(this, message, value);
    return message;
  }
  internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ResourceSpans): ResourceSpans {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* opentelemetry.proto.resource.v1.Resource resource */ 1:
          message.resource = Resource.internalBinaryRead(reader, reader.uint32(), options, message.resource);
          break;
        case /* repeated opentelemetry.proto.trace.v1.ScopeSpans scope_spans */ 2:
          message.scopeSpans.push(ScopeSpans.internalBinaryRead(reader, reader.uint32(), options));
          break;
        case /* string schema_url */ 3:
          message.schemaUrl = reader.string();
          break;
        default:
          let u = options.readUnknownField;
          if (u === 'throw') throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          let d = reader.skip(wireType);
          if (u !== false) (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
      }
    }
    return message;
  }
  internalBinaryWrite(message: ResourceSpans, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
    /* opentelemetry.proto.resource.v1.Resource resource = 1; */
    if (message.resource) Resource.internalBinaryWrite(message.resource, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
    /* repeated opentelemetry.proto.trace.v1.ScopeSpans scope_spans = 2; */
    for (let i = 0; i < message.scopeSpans.length; i++)
      ScopeSpans.internalBinaryWrite(message.scopeSpans[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
    /* string schema_url = 3; */
    if (message.schemaUrl !== '') writer.tag(3, WireType.LengthDelimited).string(message.schemaUrl);
    let u = options.writeUnknownFields;
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.trace.v1.ResourceSpans
 */
export const ResourceSpans = new ResourceSpans$Type();
// ...
export const ScopeSpans = new ScopeSpans$Type();